requirements.txt
alembic==1.2.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
decorator==4.4.1
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-pymysql==0.2.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
pbr==5.4.3
PyMySQL==0.9.3
PySelenium==1.0.2
python-dateutil==2.8.0
python-editor==1.0.4
requests==2.22.0
selenium==3.141.0
six==1.12.0
SQLAlchemy==1.3.10
sqlalchemy-migrate==0.13.0
sqlparse==0.3.0
Tempita==0.5.2
urllib3==1.25.6
visitor==0.1.3
Werkzeug==0.16.0
